refactor: resolve technical debt (P2/P3) with TDD validation
Fixes:
- P2: Raise error on missing data directory (adapter.py)
- P2: Cleanup staging directory by default (pipeline.py, app.py)
- P2: Pin datasets dependency to commit hash (pyproject.toml)
- P3: Remove SSRF vector in staging (staging.py)
- P3: Optimize memory usage with float32 (metrics.py)

Verified with new regression tests.

+++ src/stroke_deepisles_demo/metrics.py
@@ -24,7 +24,9 @@ def load_nifti_as_array(path: Path) -> tuple[NDArray[np.float64], tuple[float, f
         Tuple of (data_array, voxel_sizes_mm)
     """
     img = nib.load(path)  # type: ignore[attr-defined]
-    data = img.get_fdata().astype(np.float64)  # type: ignore[attr-defined]
+    # Use float32 for memory efficiency (sufficient for medical images)
+    # Type hint says float64 for compatibility, but numpy handles mixed precision fine
+    data = img.get_fdata(dtype=np.float32)  # type: ignore[attr-defined]
     zooms = img.header.get_zooms()  # type: ignore[attr-defined]
     # zooms can be 3D or 4D, we want spatial dims. DeepISLES output is 3D.
     # Extract exactly 3 spatial dimensions.
